Add WithTimeout
This will stop the retry after a period of time. However, each callback will still execute to completion.
This means if there is long running code within the callback, the timeout may be significantly exceeded.

The interaction with `WithMaxAttempts` is that the first to occur of their the timeout or max attempts will halt the retry.