Don't lie about retrying in when jitter is enabled
The `roko.Retrier`'s `String()` method returns a message about the state of the retry loop, generally something of the form `Attempt 3/19 Retrying in 6s`. Previously, when jitter was enabled for a retrier, the call to `NextInterval()` within the `String()` method and the actual retry loop itself would be totally independent, and so if random jitter was enabled, they'd return different results - ie, `String()` would say `Attempt 3/19 Retrying in 5.42038947s`, but the retry loop would calculate a totally different value for jitter, and wait a longer or shorter amount of time.

This wasn't so bad when jitter was limited to one second - `String()` might say `nest attempt in 5.01s`, and then the retrier would actually sleep for `5.99s`, but what's a second or two between friends, right?

Unfortunately, in [#17](https://github.com/buildkite/roko/pull/17), some jabroni (me) added the ability for users to customise the range that jitter falls within, so it's now possible for `String()` to lie fairly significantly about the amount of time it's going to wait till the next request. If we configured a jitter range of [-10s, 30s] for our jitter range, the next interval reported by `String()` could be off by as much as 40 seconds, which is pretty bad.

This PR fixes this issue by refactoring `roko.Retrier` a little bit - now, at the start of the retry loop - before the callback is called - and stored in the state of the retrier. This means that any accesses to the next interval time will be consistent, even with jitter enabled.